Rename LocalizationStore
