--- src/Sulu/Bundle/AdminBundle/Resources/js/stores/LocalizationStore/index.js
@@ -1,4 +1,0 @@
-// @flow
-import LocalizationStore from './LocalizationStore';
-
-export default LocalizationStore;
