+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/index.js
@@ -1,6 +1,6 @@
 // @flow
-import localizationStore from './LocalizationStore';
-import type {Localization} from './LocalizationStore/types';
+import localizationStore from './localizationStore';
+import type {Localization} from './localizationStore/types';
 import MultiSelectionStore from './MultiSelectionStore';
 import ResourceStore from './ResourceStore';
 import userStore from './userStore';
