+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/localizationStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import localizationStore from './localizationStore';
+
+export default localizationStore;
