+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/localizationStore/localizationStore.js
