+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/localizationStore/types.js
